KIOSK_BACKEND_V2_PYTHON/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
sqlmodel==0.0.16
greenlet==3.0.3

# AI & Orchestration
langchain>=0.1.13
langgraph>=0.0.30
langchain-groq>=0.0.1
litellm>=1.40.0

# Async & Networking
websockets==12.0
aiohttp>=3.13.0
python-dotenv==1.0.1

# Circuit Breakers & Resilience
tenacity==8.2.3
sarvamai>=0.1.0

# (Optional) Memory
# mem0ai
# zep-python
